# services/core/src/swagger/swagger.yaml
openapi: "3.0.3"
info:
  title: GlobalTransact API
  version: "2.0.0"
  description: |
    International money transfer platform API.

    All endpoints require Bearer JWT authentication unless marked as public.

    **Roles**: client | agent | admin | compliance_officer | super_admin
servers:
  - url: /api/v1
    description: Core API v1
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - BearerAuth: []
paths:
...
  /kyc/documents:
    post:
      tags: [KYC]
      summary: Upload KYC documents
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                national_id: { type: string, format: binary }
                passport: { type: string, format: binary }
                selfie: { type: string, format: binary }
                utility_bill: { type: string, format: binary }
      responses:
        201: { description: Documents submitted }
...
  /kyc/admin/applications:
    get:
      tags: [KYC]
      summary: List all KYC applications (admin/compliance)
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [PENDING, UNDER_REVIEW, APPROVED, REJECTED, REQUIRES_DOCS] }
      responses:
        200: { description: Applications list }
  /kyc/admin/applications/{id}/review:
    post:
      tags: [KYC]
      summary: Review a KYC application
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision: { type: string, enum: [APPROVED, REJECTED, REQUIRES_DOCS, UNDER_REVIEW] }
                reviewNotes: { type: string }
                rejectionReason: { type: string }
      responses:
        200: { description: Review saved }
...
  /otp/request:
    post:
      tags: [OTP]
      summary: Request an OTP code
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [purpose]
              properties:
                purpose: { type: string, enum: [transaction_confirm, login_2fa, phone_verify] }
                reference: { type: string }
      responses:
        201: { description: OTP sent via SMS and email }
  /otp/verify:
    post:
      tags: [OTP]
      summary: Verify an OTP code
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [otpId, code, purpose]
              properties:
                otpId: { type: string, format: uuid }
                code: { type: string, minLength: 6, maxLength: 6 }
                purpose: { type: string }
      responses:
        200: { description: OTP valid }
        400: { description: Invalid or expired OTP }
...
  /reports/request:
    post:
      tags: [Reports]
      summary: Queue a new report generation
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [type]
              properties:
                type: { type: string, enum: [transaction, commission, forex_gain_loss, compliance, audit, settlement, revenue] }
                dateFrom: { type: string, format: date-time }
                dateTo: { type: string, format: date-time }
                corridor: { type: string }
      responses:
        202: { description: Report generation queued }
  /reports/{id}/download:
    get:
      tags: [Reports]
      summary: Download a completed report
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        200:
          description: Report data (JSON)
          content:
            application/json:
              schema: { type: object }
...
  /super/config/{key}:
    put:
      tags: [Super Admin]
      summary: Update a system config value
      parameters:
        - in: path
          name: key
          required: true
          schema: { type: string }
      responses:
        200: { description: Config updated }
...
  /super/users/{id}/role:
    put:
      tags: [Super Admin]
      summary: Change a user's role
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        200: { description: Role updated }
  /super/users/{id}/status:
    put:
      tags: [Super Admin]
      summary: Suspend or reactivate a user
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        200: { description: Status updated }
...